pyproject.toml
[project]
name = "peek-challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.171.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "accelerate>=1.4.0",
    "datasets>=3.3.2",
    "faiss-cpu>=1.10.0",
    "fastapi>=0.115.11",
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.18",
    "langchain-core>=0.3.40",
    "langchain-ollama>=0.2.3",
    "pdfplumber>=0.11.5",
    "streamlit>=1.42.2",
    "transformers>=4.49.0",
    "uvicorn>=0.34.0",
]
